서버 베이스, 자동화 작업 불필요한 파일들 제거

+++ Server/Base/Component.h
@@ -23,7 +23,7 @@ public:\
 	name() : CComponent() { Init(); } \
 	virtual ~name() {} \
 	static std::string	GetName() { return #name; }\
-	static int			GetHash() { return std::hash<std::string>()(#name); } \
+	static int			GetHash() { return (int)std::hash<std::string>()(#name); } \
 	static CComponent* GetClone() \
 	{ \
 		return new name(); \
